Update patch set 1
Patch Set 1:

You'll need the bt-next branches of bluedroid, Bluetooth, and f/b to test this. It's not required to enable the BLUETOOTH_QCOM_SW bit in Bluedroid, that's only for actual QC chips.

The BluetoothExt contains the upstream MAP implementation, FTP, BTC, and also SAP (untested, needs an extra framework patch and possibly RIL support).

I reverted our MAP support from the Bluetooth package, but we have a few bugfixes that may still be applicable to the new implementation.  MAP seems to be working great for me now, could you guys give it a spin?

Patch-set: 1