Update patch set 1
Patch Set 1:

Not to hijack this conversation, but your findings on BTC match mine. If you turn on the debug logging on the BTC service, it constantly complains nothing is connecting to its socket.

In my tests with apexqtmo, what I'm seeing is Bluedroid is apparently working correctly, but nothing is passing WLAN_NL_MSG_BTC messages to the kernel for the prima driver to pay attention to (grep the prima code for this phrase).

Patch-set: 1