Update patch set 1
Patch Set 1:

I spoke with Xplod at length about this. I think Focal is a great piece of work with huge potential, but there's now this perception that CM is trying to steal his hard work and run with it. I would rather see him continue this as a standalone project (which will no doubt be successful) rather than continue to generate drama in the community.

Patch-set: 1