Update patch set 4
Patch Set 4: Code-Review-1

Why would we programatically create a "stubbed" apn if we already have an apn in the list? Why remove the apn to create one programatically instead? the stubbed apn is a fallback for is no apn is found at all. were full circle using the stubbed apn again because we removed the actual apn from the list.

Patch-set: 4