Update patch set 4
Patch Set 4: Code-Review+1

We must provide the default behavior that is not a default apn. 
 the problem with your patch, you are trying to create a connection, on the nexus devices, 
it tries to connect to 3g. it says an apn is found, modem 
searches and said,  not connectable due to the effect, 
there is no apn to connect to  and conflicts because the 
apn are the same  and probably won't connect to that apn. 

i fixed the provisioning bug and now 3g is back with  stubbed apn.

did you even consider Mike Kasick (mkasick) comment and considered it to be a problem (re-- conflicting apn and opeerator numeric issue). We are having problem on jf and d2 when it comes to connect lte network and ehrpd. 

It is better to create a programaticble. it checks if it is cdma and apn list is empty, just like caf patch does it.

Patch-set: 4