Update patch set 4
Patch Set 4:

The APN isn't found and that's why it creates a stub. That is the behavior we want yes (in the event no APN is found). It's a fallback. Here we have an APN being defined, found, and use. All of which it should be. But because the telephone stack isn't prioritizing the APN correctly you are removing it and instead creating it programatically. 

The issue here is it finds an APN and doesn't have a sense of "preferred" APN. The real issue seems to be in selection of a preferred APN vs. the fact we're not creating the same APN programatically. The "CAF way" is for OMA devices where APNs may not be on the device. There is also a API to retrieve the OMA provisioning info as well. I wouldn't call our current way the "CAF way". Since all we do is create an APN programatically if one isn't found with reliance on other found APNs to generate the configuration.

Ever wonder why our there isn't this same issue on Verizon and such that have default APNs as well as eHRPD, DUN, LTE, and FOTA? 

So yes, I understand the issue. I just don't think removing the APN is the right  solution. I think making the other APN "preferred" is closer.

Patch-set: 4