Update patch set 4
Patch Set 4:

all those posts say is that data handoffs work but messaging is now broken and so on. it "fixes" the apn selection yes. but not by actually selecting the correct apn based on RAT type. instead by just not having an apn to select other then the eHRPD or LTE apn. which forces it to fallback on the "stubbed" apn which breaks MMS and such. As i said, it's not that the patch doesn't "fix" data handoffs. its that you are only bandaiding the issue and subsequently creating another issue (broken mms).

I'm working on a patch to handle apn selection based on RAT type. It should resolve the issue and allow for apns to be defined and selected correctly based on RAT in use.

Patch-set: 4