Update patch set 8
Patch Set 8: Code-Review-1

the apn isn't the issue. it's the logic involved in apn selection based on RAT type is the issue. If the device is attempting to acquire a EvDo connection it is okay for it to select the APN for that _and_ us to have a default/hipri/preferred apn defined for it here. The MCC MNC matching all other RAT type defined APNs isnt the issue either. 

APN selection should be based on two things combined. One, the MCC MNC must match the carrier. Two, the RAT type must match that of the connection we are attempting to connect to. obviously there is issue with the later half of APN selection and when attempting to connect to a eHRPD (and probably LTE) connection. the logic for updating the APN to that of an apn with a matching bearer of the connection we are attempting to connect to has a flaw.

again, removing the apn is not the solution. fixing the logic when selecting apns is the solution. removing it does nothing but force us to fall back on a stub. which also requires us to switch from the stub to the mms apn when sending mms, and back to the stub again after. we dont need the stub. we are defining it here. fixing the logic so that when attempting to connect to RAT type 13 or 14 the APN that is selected needs the bearer type of 13 or 14 to match it. instead of considering 0 bearer to be acceptable for a connection requiring bearer of 13 or 14.

Patch-set: 8