Update patch set 3
Patch Set 3: Code-Review+1

Just to explain this, when on a network extender, an operator numeric of 310006 can be seen in radio logs on VZW devices, and these devices fall back to stubbed APN when an actual APN isn't found which can sometimes cause issues if the fallback fails etc...  This adds APN's for the 310006 numeric to satisfy the request.

This has been extensively tested on jfltevzw and d2vzw and works exactly as it should and the device no longer relies on stubbed APN for data/mms on Verizon Network Extenders.

Patch-set: 3
Label: Code-Review=+1