Update patch set 5
Patch Set 5:

That's not my point, we do want a single apns file. The pre-existing apns-conf.xml was already used by all (or should be) GSM models. We should not be bringing in additional GSM configs "out of nowhere" and risk breaking the (majority of) GSM models that had good configs

Patch-set: 5