Update patch set 6
Patch Set 6: Code-Review+1

This sticks to strictly moving the cdma apns over to the main file. The default sprint and verizon config that was in the gsm apns was removed in favor of the working ones from the cdma apns file. Tested on verizon and straight talk with no issues.

Patch-set: 6