Update patch set 1
Patch Set 1: Code-Review-1

Well, it's not needed as long as the dex file is compatible with dex2oat which runs at boot. It can be done perfectly well without adding odex files.

With that said, I agree they probably include them as a workaround for ART. I wonder how that would work though since odex files are tied to the specific versions of the jars in the CLASSPATH of the device they are generated on. Removing that dependency is the reason deodexing is used in the first place.

Patch-set: 1
Label: Code-Review=-1