Update patch set 1
Patch Set 1:

Yeah, we probably need to do something like this; when the SIM card doesn't export a SPN, it's pulled from the radio firmware; radio firmware data is often outdated, leading to either wrong text or just the MCC/MNC, especially on hardware that's been "abandoned" by the OEM and white-label carrier-branded stuff that wasn't meant to run on other carriers.

That being said... If we overload the SPNs, let's try to make sure the values are correct. Starting with not replacing the original values we had there already.

Patch-set: 1