Update patch set 1
Patch Set 1: Code-Review-1

This is switching the Sprint APNs from the correct MCC values to the Test MCCs? This should not be submitted. If you take a look at the apns-full-conf.xml in device/sample here:
https://github.com/CyanogenMod/android_device_sample/blob/cm-11.0/etc/apns-full-conf.xml

You can see that the correct MCC and MNC are 310 and 120. The test file you pulled from (i'm assuming in device/samsung/toroplus) is out of date and should not be used. The file I provided is the one used in stock Android builds for hammerhead (Nexus 5) and is the correct format.

Patch-set: 1
Label: Code-Review=-1