Update patch set 1
Patch Set 1: Code-Review+1

arcee, I'm purposely not adding a boatload of reviewers because I've seen how contentious similar ideas have been in the past. If you ask, I will abandon this. Otherwise, I'll leave it to you to either add more reviewers or approve it outright.

Obviously, I like this idea, and I'd even like to pitch it as a "CM recognized" way to custom tag unofficial builds. Right now, if multiple people provide unofficials for the same device the output versions can be identical for differing releases. The differences can come from time of day for the compilation, backporting commits, or testing unapproved commits. This information could be useful to maintainers who monitor forums to see how a device is performing. These tags (which "modders" will probably like to use) could help maintainers identify whether reported issues are really worth looking at. In this sense, I'd go so far as to suggest advertising the ability to tag releases on the wiki.

Ok, my cards are on the table. I'll leave it to you.

Patch-set: 1