Update patch set 5
Patch Set 5:

Not to sound overly harsh (and I don't mean for it to be): Why should mainline CM be affected/impacted for legacy devices that will never see official support?

Now, don't get me wrong, I +1'd the 10.2 version of this because of the intent of the patch and the impending finality of that code branch - but I struggle to see why 11.0 mainline should have patches solely for the sake of legacy/unofficial devices. 

I think therein lies the point Ricardo is making. Where is the line where this patch is useful to "The CyanogenMod Project" versus the home builder. 

(Yes, I fully realize this is a fairly benign change, but its still a discussion worth having).

Patch-set: 5
Label: Code-Review=0