Update patch set 5
Patch Set 5:

I think all of the legacy changes present in mainline CM code are present for only 1 reason: make the life of independent developers easier.

Even if the device will never get official support again, it's a lot easier for us to have these patches in CM instead of forking every repository we need to modify and then having to keep up with the new changes CM is pushing.

Patch-set: 5