Update patch set 5
Patch Set 5:

@Michael Bestas

Sorry, I don't think I explained that very well. I'm saying that if you are on an unofficial build of cm-11, and there are official builds available of an older version of CM for the device (like cm-10.2), that they don't show up, and if the device gets official cm-11 support (or whatever version you're on) eventually, then those builds show up in CMUpdater so that users can update to them, but can't downgrade on accident to an older "official" version.

Patch-set: 5