Update patch set 1
Patch Set 1:

Nick: Yeah, /system labeling is on its way out. The problem with user data is that we're not upgrading from unlabeled, we're upgrading from different labels (and we were using upstream seandroid's more complex stuff with per-app contexts and the like. We've now switched to AOSP).
I'm afraid we do need the restorecon, unless you can think of something better to clear away the existing metadata ;)

Patch-set: 1