Update patch set 5
Patch Set 5:

ciwrl: I can't speak for anyone else (though there seems to be interest), but we're looking at this for encore, which has 440.5 MB on /system and needs some serious squeezing to get in under that limit.

I have no particular objection to keeping VoicePlus, but neither do I see it as important enough to be kept on devices that are already squeezed enough for space to want to use the mini build variant.

Marin: comparing encore builds with and without these changes, I see a savings of 4868 KB, at least 2 MB of which comes from losing speech recognition stuff that goes away with VoiceDialer.

Patch-set: 5