Update patch set 1
Patch Set 1:

I haven't had time to verify this myself, but if we change the import order, the problem may be resolved. The stock init.rc imports in this order:

import /init.environ.rc
import /init.usb.rc
import /init.${ro.hardware}.rc
import /init.trace.rc
import /init.carrier.rc
import /init.cm.rc

If the order matters, the init.cm.rc is parsed last (and the ksm stuff is configured in the "on boot" block, which is the final phase), so it overwrites values set in the $ro.hardware config.

I suppose we could workaround this by updating the values in the per-device configs via "on property:sys.boot_completed=1". I'll give that a try.

Patch-set: 1