Update patch set 1
Patch Set 1:

EXEC is sync, services are async. That means when you call exec, it waits the init.d scripts, then core and main services are run.

The problem with the current implementation is that init.d scripts are run too late and that breaks all apps2sd applications & scripts. By breaking I mean that on every reboot, dalvik-cache is regenerated for the apps that are linked/moved to SD.

Patch-set: 1