Update patch set 1
Patch Set 1:

Yes, but this service in particular isn't executed by class_start. It's directly invoked at the end of the post-fs-data block, before "on boot" (where this is now being placed) even starts. It can be switched to an exec instead of an async service start, but I wouldn't take it out of post-fs-data

Patch-set: 1