Update patch set 1
Patch Set 1: Code-Review-1

This cannot be added to a sub-init.rc file. I have tested this before, and even if you use EXEC, it will be executed async compared to the content of the main init.rc file. The changes in the 57603 commit is the correct way of doing this. The EXEC needs to be placed within onBoot in the main init.rc file. Otherwise you might as well be using a service.

Patch-set: 1