Update patch set 6
Patch Set 6: Code-Review+1

This patchset overwrite the 'default' eri.xml file with the VZW one in the XML folder.  Not sure if that will cause other problems, but it shows that something is not pulling the correct eri for the mcc/mnc on Verizon Wireless on d2vzw for some users.

This fixes the "Roaming Indicator Off" for my d2lte(d2vzw).

Patch-set: 6