Update patch set 11
Patch Set 11:

It looked like about half of the VZW devices we support had differing roaming indicator settings. I wish this was consistent, but Verizon hasn't made it that way. Since the vendor/cm overlays are meant for truly universal settings, this didn't belong here. I'm all for it being unified where it can, but it just isn't a universal solution.

Patch-set: 11