Update patch set 1
Patch Set 1:

Danny, committing the overlays to devices that need them is the option. You really have to dispel the thought that if an overlay is provided for a specific mcc-mnc, it must be universal. While I'm not necessarily against the idea of some of these spn-specific overlays being in vendor/cm, finding out which overlays are truly device independent is even more time cumbersome than just adding them to devices which may use them. Putting the overlays in device-common repositories should help relieve some of the pain, though.

Patch-set: 1