Update patch set 2
Patch Set 2:

This would make sense here, except it makes every device claim to softbank that it's a nexus5, and that's just wrong. We reverted the verizon bits, because verizon is verizon, but let's not assume every carrier is as "speshul"

If there's a softer device-neutral value that can be set here, I'm all for putting it as a global overlay. This particular patch shouldn't be placed anywhere other than the hammerhead.

Patch-set: 2