Update patch set 4
Patch Set 4: -Code-Review

Steaven, do a radio logcat and look for one of these 2 values: 
Set CDMA Roaming Indicator to: x (where "x" is a number value)
or
mRoamingIndicator = x (where "x" is a number value)
Report back your value. Also, this is really not the appropriate place for a general question not related to the patch. Your question is an ERI related issue...  :)

Patch-set: 4