Update patch set 1
Patch Set 1:

Tony, I believe the idea is benchmarks should be set by the user. It should be for them to see what they are getting for their current choice. Intel/AMD are not automatically overclocking/pushing their cpu's everytime someone runs a benchmark. If the user choose to overclock and see what they gets them then so be it but choosing for them in a bit odd to me.

Patch-set: 1