Update patch set 1
Patch Set 1: Code-Review-1

A counterargument would be that the utility of benchmarks and other unrelated applications is not the operating system's responsibility, that it behooves the developers of these applications to make their products useful. While we could request the desired changes, there are many benchmarking apps that would need to be updated, and it's difficult to coordinate all the moving parts. Although special-casing these applications on our end might feel sub-optimal, it produces results that (imho) are less misleading and more useful in the interim. These are the alternatives:

1. Keep the previous commit: benchmarks approximate the upper limits of performance for CyanogenMod devices, and are useful for making general performance comparisons
2. Merge this commit reverting the former: benchmarks approximate the average-case performance for CyanogenMod devices, averaged across benchmarking conditions rather than daily-use conditions, and are pretty much useless for pretty much anything

Patch-set: 1
Label: Code-Review=-1