Update patch set 1
Patch Set 1: Code-Review+1

I don't know if this is the proper way of conditionally building it, but it works when I do my unofficial builds.

I was requested to do this as an user reported that CMUpdater was running all the time & since there's no use for it, why waste it.

Patch-set: 1
Label: Code-Review=+1