Update patch set 1
Patch Set 1:

I need another chance to play with this, as I was getting some issues with a pre-release edition of the OnePlus. I merged the tracking just so people didnt have to manually pick that one all the time ;)

This will go in, I just want to be sure what I saw before wasn't a persistent issue.

Patch-set: 1