Update patch set 2
Patch Set 2:

I think the only scenario I envision this making a difference is 3G or 1x networks, perhaps Verizon's LTE network is the only one which accepts apn==VZWAPP (i.e. notice that type==mms is technically covered for this spn, but the question is whether there are networks for which only apn==internet is acceptable).

I would not suggest referencing other spns as justification for why 311480 should include this line. Most CM vzw devices use 311480 by default and so this spn gets the most attention.

I'll test this out on m7vzw to see if MMS still sends on LTE, 3G and 1x networks.

Patch-set: 2