Update patch set 2
Patch Set 2: -Code-Review

I see.

It could be that the 20140911-20140916 issue (NIGHTLIES-178) disabled the VZWAPP cbs APN.

In which case, by using the internet/VZWINTERNET apn for mms, users such as Disco Orange were able to keep sending texts.

If this is the issue, and cbs works again for everyone, then my suggested patch is indeed useless.

In fact, perhaps it could be done the other way, and the other 15 Verizon internet mms APNs could be deleted.

Patch-set: 2