Update patch set 1
Patch Set 1:

Isn't it a problem with the apn code? Why isn't the phone trying the other default apn if the first one fails?

If indeed the default type should be removed, then it should probably also be removed from the FreedomPop and Virgin Mobile apns. Might as well reorganize the apns so they are in alphabetical order).

Patch-set: 1