Update patch set 2
Patch Set 2:

It was recommended to use the APN file extensions by a google engineer, unfortunately it's not documented. Also in bare Kitkat on the Nexus4 it seems the IMSI match is not yet taken fully into account (our SIM also shows other 20408 APN's while in 20408 mode), but there are others in the apns-full-conf.xml as well. Also in CM11 you can check the APN configuration screen: here you can fill in "MVNO type" and "MVNO value".

Pls check:
https://android.googlesource.com/device/sample/+/master/etc/apns-full-conf.xml

Patch-set: 2