Update patch set 1
Patch Set 1:

Joey, I would normally agree, but this is required to build CM.  It could be made optional -- and I have submitted a patch to do just that in the past -- but apparently that's not desirable either.

I think the long term solution is to develop a CM terminal emulator.  But until that happens, I'd like to prevent new developers from tripping over this issue one way or the other.

Regarding reproducing builds, the build manifest is saved in each build under /system/etc.  This can be used to recreate a code-identical build, but it will not be byte-identical due to various external factors.

Patch-set: 1