Update patch set 1
Patch Set 1:

Well, depends on what you mean by required, but I'm not saying it should be removed entirely.  I see what you mean, that the makefiles depend on it, as I experienced that when doing my first build.  Now I put those files in PRODUCT_COPY_FILES_OVERRIDES so it's not required.

I agree it would be nice to help new builders.  Can you just ensure a more helpful error message is shown instead?  ("Term.apk not found, please run get-prebuilts to have it downloaded, then rerun the build command.")  Isn't the same type of error going to happen if blobs haven't been pulled?  So why not just put an instruction to run get-prebuilts, alongside the instructions to run extract-files.sh (actually, it already seems to be there.)

A comment in the current get-prebuilts file seems to suggest the only reason is so that it can updated from the Play Store.  Why is that important, but only for this one system app?  Or why not build it in-tree with a non-colliding name, and let people who need updates install that one from the Play Store?  Maybe a branding/licensing issue? -- I understand AOSP has a terminal emulator; why not build it instead?  Or have an open-source, CM-built component that prompts on first boot to install from Play Store (if gapps are present.)

I know about the build manifests, and they are handy, as I use them to recreate M releases.  I'm talking about building this Term app in the CM tree, in such a way that it's exactly what is released on Github, and putting that signature into the built apk.  I recognize this could be very difficult to achieve, especially without silly build system requirements.  (It would be cool to have this auditability for the whole CM release, but would be even more difficult.)

I recognize this may seem a bit silly to be arguing about, in a project that requires closed-source blobs to boot, let alone use half the hardware, on most devices -- I don't want that to be an excuse to do it more than is necessary to have a minimally usable system.  (Whether a terminal emulator is required for that is a bit debatable, but again, why do we need this particular one, and why does it need to be updateable from Play Store?)

Patch-set: 1