Update patch set 1
Patch Set 1:

P.S. I'm not particularly arguing for the alternate ways I talked about, just wondering what are the requirements and preferences that are driving this decision.  I think it's an OK compromise as it is: it works if you follow all the build instructions on the wiki, but it's transparent to advanced users.

As an implementation comment: I would have the makefile invoke get-prebuilts, or something to not duplicate those shell commands in the two places (for maintainability, clarity, etc.)

Patch-set: 1