Update patch set 1
Patch Set 1:

I have also found it strange that we include this as a prebuilt apk, when the same thing is available from the play store.

can't we change get-prebuilts to generate a .mk file, so we can use "call inherit-product-if-exists" so that builds don't fail if we haven't run the script?

Patch-set: 1