Update patch set 1
Patch Set 1: Code-Review+1

What we have (before this patch) is the same as AOSP, and isn't working.
This change reflects what supposedly is Samsung's configuration for this APN (mcc 405, mnc 01). 
I don't have a way to test this, so I can't be sure if this fix is correct - only that the current configuration isn't.

Patch-set: 1
Label: Code-Review=+1