Update patch set 2
Patch Set 2: Code-Review-2

@Martin- Design is a highly subjective thing
 But in this case, it's literally a copy of someone else's design with the name changed. This looks good and all, but it doesn't feel original to me hence my thumbs down. We have an entire team of designers who are totally amazeballs, and we've also had great results as a community as well (all the past anims, Cid, etc).  The first thing you see when you fire up CM should be fresh and speak to our originality. We are creators here and I would be a shitty leader if I was happy just copying someone else.

Also, it's hard to test every legacy device when working in the more fragile layers (like Stagefright). I try to get as much testing in as possible but sometimes the only way to make an omelet is to break a few eggs. It's *why* we have nightlies! That said, I'm one of the only people working in this layer and it's difficult to get reviewers, any assistance is more than welcome.

Patch-set: 2