Update patch set 2
Patch Set 2:

Yes, it will fall back to IPv4.  But that's not what we want.  Using the "telenor.mobil" IPV6 APN allow the older devices to establish IPv6 connections, with NAT64 and DNS64 support from the network.  This gives the devices full dual stack access when combined with the 464XLAT translation in Android.

Falling back to IPv4 only is a step back.  Please use the "telenor.mobil" APN for any device without IPV4V6 support.

But do note that the old MMS proxy setting obviously was wrong.  It should be "mms-proxy.telenor.no".  Using an IPv4 literal in an IPV6 APN does not make much sense...

Patch-set: 2