Update patch set 8
Patch Set 8:

I tried prepending the needed values instead of appending them when setting the actual variables, but the issue was that the inheritance system started all of the product variables empty and then resolved them afterward in the build process - so in my makefile was prepending to an empty variable, instead of appending to one.

To address Steve's comments about some of the other components in vendor/extra requiring values that are set up in vendor/cm, I created an inherit function that will prepend values instead of appending them - basically taking the suggested approach above and applying it to the inheritance system.

The other possibility was to create a special, "master override" overlay variable that would be referenced directly by build/core/product.mk and prepended to *everything* else.. but I can foresee that immediately running into this exact same problem the next time someone tries to prioritize something, so I really didn't want to go that route.

Patch-set: 8