Update patch set 2
Patch Set 2:

The "telenor.mobil" APN is definitely not working on my old device. It doesn't connect at all. And since IPV6 tethering is currently somewhat broken in CM11, I think we should revert to using the old APN (IPV4 only).

For CM12, where IPV6 tethering works fine, I don't see why the new APN should be any problem, if it provides IPV6 for those devices that support it, with a fall back to IPV4 for those that doesn't. Maybe I'm wrong, and I'm no expert on this, but isn't that what IPV4V6 means? That both protocols are supported?

Afaik, configuring this per device is not possible, as we have one common APN list which all devices inherit from.

Patch-set: 2