Update patch set 2
Patch Set 2:

It makes me nervous to upload a new patchset out of fear for upsetting some folks. There are people who haven't uploaded changes to devices they maintain in over a year, but are still around in some capacity.

It might be easier if we could get a general rule to blindly apply: if no patches in 12mo, name removed from device.

Patch-set: 2