Update patch set 9
Patch Set 9:

Tablets typically have a natural orientation that is landscape.  Picassowifi fits this category and the animation looks rather silly on it (huge black bars on the side).

This could, I believe, easily be fixed by changing the boot animation:

a) The desc.txt file contains an optional background color per part.  This is not currently specified and should be set to the background color used in that part.

b) The images seem to have a slight border around them (when I hacked up the background color, there was a grey line down the side where the bar used to be).  This should be removed so that the entire background is the background color.

Patch-set: 9
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0