Update patch set 9
Patch Set 9:

crpalmer: I like the idea, but it would still be weird going from a part with a white background to a card animating out (with a black background). Maybe another part is needed to transition from the white background to the black with the card. Not sure how that would be possible,though.

Patch-set: 9