Update patch set 1
Patch Set 1:

I mean, we can - but should we? We don't exactly have info on user expectations of this being here, outside of reports of it not resolving. 

Further, going with this approach as opposed to Danesh's prior fix http://review.cyanogenmod.org/#/c/67640/ to prevent losing this on upstream merge again.

Patch-set: 1