Update patch set 1
Patch Set 1:

The key issue with the white version, as I see it, is that it's shown at a time when the screen is typically at max brightness since settings are not yet applied.  So not a problem with white, per se, it's just not a good time to be displaying a lot of it.

Patch-set: 1