Update patch set 1
Patch Set 1: Code-Review+1

There seems to be a bug in how the default wallpaper is applied though. Here's how it looks like:

http://imgur.com/aYUReVw,oxL9Rsq#0

If I apply the same image from CM Wallpapers category, it looks like this:

http://imgur.com/aYUReVw,oxL9Rsq#1

The last one is correct as it lines up with the edges of the original image.

Patch-set: 1
Label: Code-Review=+1