Update patch set 4
Patch Set 4: Code-Review+1

Unless it's preferred that I move this to my device -- which I'd understand since not all devices use UMS -- I'm in favor of this. As much as I'd like to narrow it down to a more specific label, I just couldn't make that happen (see previous comment).

Patch-set: 4
Label: Code-Review=+1