Update patch set 4
Patch Set 4:

There is no official VZW model of nexus 6 shipping so they all default to load the GSM fsg image to the modem.  If you do 'fastboot oem config carrier vzw' it will load the cdma fsg image to the modem, which prevents the issue of being stuck on 3g or 1x.  You need to flash a nightly with the modem image in it after you do this to reset the modem.  Even before this commit I had lots of issues with LTE until I set the carrier.

Patch-set: 4