Update patch set 1
Patch Set 1:

A better fix would be to put bearerdata= 14 and 13 on those affected apn numeric. because it seem that some people are having issues connecting to lte. So there are some apn are being used for lte itself so set the bearer to 13 and 14

Patch-set: 1
Reviewer: Shareef Ali <883@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>