Update patch set 1
Patch Set 1:

(3 comments)

This sounds very useful, and has been supported since a while too, nice.

Allowing to enable this for specific modules (chromium is the first thing that comes to mind) might be useful too.

Also, could something like this also be done for java libraries? I understand that hooks for those aren't present, but those would be trivial to add compared to the actual implementation which is done here.

Patch-set: 1