Update patch set 1
Patch Set 1:

(2 comments)

Replies inline.

This will cache chromium and many, many other modules.  I don't see much usefulness in allowing only specific modules to be cached.  If it works and it saves time, it should work for all modules.

Tests would need to be done to determine if caching is useful for java libraries.  Conventional wisdom says that java compilation is fast enough that caching generally doesn't help.  But on a library level, it may help some.  Feel free to make a test and share the results.

Patch-set: 1