Update patch set 1
Patch Set 1:

A little explanation: this is supposed to avoid this (experienced on bacon, device never goes deep sleep):

https://docs.google.com/file/d/0B8BdLIPvFvR5NXFQY2Z2Z3RxaDA/edit?usp=docslist_api

More than 8 minutes of wakelock on less than 9 minutes of screen off.
Many people in the web suggest to disable all the 3 receivers of SystemUpdateService in both Google frameworks services and Google play services and this is the only one that's not disabled now.
At the moment I'm not able to test this change directly, but I disabled this receiver using a service manager app and the wakelock disappeared. I'll test this as soon as possible, however it would be great if someone could try this meanwhile.

Patch-set: 1