Update patch set 1
Patch Set 1:

I know very little about this topic or how it works, but I've been seeing this wake lock with hammerheadcaf, and flo nightlies. Enabling the SystemUpdateService and leaving it enabled clears the wake lock for me. (Until reboot). 

This would suggest to me that it should be left on. (But as I say, I'm not a dev, and know very little about this).

Patch-set: 1