Update patch set 1
Patch Set 1:

I've put a little more thought into how I should have worded this. What I mean, is that with SystemUpdateService disabled on boot, it seems that something is calling on Google Play Services to check whether an update is needed, whether this be for the Play Store, or the Play Services itself, or an OTA - and as this service is disabled, it keeps polling to check and receiving no answer. Hence the wake lock.

Would Play Services not do a system check on calling this service and realise "this is a custom ROM, no action needed" and go back to sleep with regards to OTA updates? If this is indeed also used for Play Services and Play Store updates, then should this service not by rights be started on boot, and allowed to manage itself, rather than being forcefully disabled?

Patch-set: 1