Update patch set 1
Patch Set 1:

I played a little with these services but I didn't get anything good. I managed to block the wakelock by adding a check in the wakelock creation routine in PowerManager, but I think this isn't a really good solution and moreover it's making google play services to crash once when the system boot is completed.

Patch-set: 1