Update patch set 1
Patch Set 1:

I'm a user, not a developer, but this is my experience with this bug.

I boot the phone, if reciever SystemUpdateService$SecretCodeReceiver is on and i disable it, the phone goes to deep sleep.

I boot the phone, if reciever SystemUpdateService$SecretCodeReceiver is off and i enable it, the phone goes to deep sleep.

I understand that the point is, change the state just boot, to dismiss wakelock - deep sleep.

I hope this help.

Patch-set: 1