Update patch set 2
Patch Set 2:

I can confirm this issue when cdma switches mode from lte to  legacy cdma. It reqiuire me to restart or manually set the apn. The Apn set a state. the boost apn is setting it as the dafault fallback case. but there is no way to reset  the case without changing the numeric.  sprintt has  only one apn numerics which is shared by lte and legacy cdma.

Patch-set: 2