Update patch set 2
Patch Set 2:

And SystemUpdateService was shown as enabled in 'adb shell dumpsys package' when you had the wakelock?

As said, re-enabling all components wholesale is not going to happen, as that will cause download loops of OTA updates on Nexus devices again. If the current solution doesn't work, a proper root cause analysis needs to happen.

Patch-set: 2