Update patch set 3
Patch Set 3:

Henry, just FYI, the matching is the first hit of typer="default" based on MCC/MNC pairs. Since these all share that mcc/mnc pair, the top most item on the list 'wins' the race, regardless of the fact of should it work or not. 

Unfortunately this is a limitation of how Android handles the behavior. At the least it does a good job of pulling all available matches and letting the user switch easily.

Patch-set: 3