Update patch set 3
Patch Set 3:

Sorry, PS2 doesn't work. I originally loaded this page in hopes no one would have reviewed it yet/changed PSes, so I was on this page with only PS1 and have verified that PS1 works. The reordering in PS2 to the original file does not change the capability of Rogers getting data again. The original PS was meant to remove the pay-as-you-go carriers as they conflicted with Rogers, and revert the APN logic for Rogers back to the previous Rogers APN as it had major issues. Readding them but reordering doesn't make sense. -1 on Nexus 6.

Patch-set: 3