Update patch set 3
Patch Set 3:

Sorry, but again, that makes no sense to me. 

The XML list is parsed in order of MCC/MNC pairs then loaded to the telephony provider db. 

The telephony provider then uses the db to determine which APN to connect to (it can only connect to one at a time). On this process, the first APN to match MCC, MNC and type having 'default', it will join. 

The reorder of the pre-pay/pay2go APNs below the 'proper' Rogers ones means that it should never even select those unless a user manually changes it to that (Rogers LTE apn being the first one to be hit, will be the one defaulted to). 

The only way to override this logic, and possibly resolve to those other APNs is if your device tree is manually configuring an override (example USA AT&T - https://github.com/CyanogenMod/android_vendor_cm/blob/cm-12.1/overlay/common/packages/providers/TelephonyProvider/res/values-mcc310-mnc410/config.xml) 

Are you certain that with PS3, hitting "Reset to default" in the APN overflow menu is resulting in selection of the prepay/pay2go APNs?

Patch-set: 3