Update patch set 1
Patch Set 1:

As soon as my previous commit showed up in the changelog, I was informed that Noverca is out of business. I was also asked to fix Vodafone IT, because by default "Vodafone WEB" (now renamed to "Vodafone Tethering") is selected by default and Vodafone charges 4€ as soon as that APN. So yeah, that's not great default choice.

Patch-set: 1