Update patch set 1
Patch Set 1:

This solves a problem with USB Mass Storage that presents itself particularly with the DriveDroid app. DriveDroid instructs the kernel to host a file over USB Mass Storage, but the kernel cannot read the file, because of SE Linux policies. This started to happen in CM-13.0.
This causes the kernel to host an empty file, while it responds to the app it has hosted the file successfully.

http://sholland.org/thoughts.html for more information about the issue.

Many thanks to invisiblek, [Saint] and thompatry for the help on IRC.

Patch-set: 1